Support escape of next key press with Control-V
Add support to escape the next key press by pressing Control-V.

This is useful, for example, when a context helper is in use, which dedicates
the '?' key to triggering a call to the context help function. In this case,
there is no way to get a '?' character.

With this change one can escape the '?' key by pressing Control-V before
pressing the '?' key  to obtain a '?' character.

This can also be used to get other special keys, like <Tab>.